requirements.txt
streamlit
ultralytics
easyocr
opencv-python-headless
numpy
pillow
torch